requirements.txt
Flask
geopy
flask-cors
pyswisseph
requests
